split compare functionality, upgrade numpy to 2.2.3
compare is nwo modularized to several files, making it easier to maintain.
