+++ requirements.txt
@@ -5,4 +5,4 @@ pytest==8.3.4
 deprecation==2.1.0
 
 # https://stackoverflow.com/questions/78634235/numpy-dtype-size-changed-may-indicate-binary-incompatibility-expected-96-from
-numpy==2.2.2
+numpy==2.2.3
